Changed satellite table to use react-query to handle fetching data. TODO: Implement useMutate and pagination with this and then take over the world
