+++ website/src/components/SatelliteTable/data-table.tsx
@@ -11,6 +11,8 @@ import {
   useReactTable,
 } from "@tanstack/react-table";
 
+
+
 import {
   Table,
   TableBody,
