+++ website/src/main.tsx
@@ -7,7 +7,7 @@ import { BrowserRouter } from 'react-router-dom';
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <BrowserRouter>
-    <App />
+          <App />
     </BrowserRouter>
   </React.StrictMode>,
 )
