+++ website/src/types/DataTableProps.ts
@@ -9,3 +9,4 @@ export interface DataTableProps<TData, TValue> {
   // this is a function that will be called when the data changes in the table
   onChangedData: (data: TData[]) => void;
 }
+
